LPS-185852 Moves from portal kernel
