+++ modules/apps/dynamic-data-lists/dynamic-data-lists-web/src/main/java/com/liferay/dynamic/data/lists/web/internal/asset/DDLRecordSetClassType.java
@@ -15,7 +15,7 @@
 package com.liferay.dynamic.data.lists.web.internal.asset;
 
 import com.liferay.asset.kernel.model.ClassTypeField;
-import com.liferay.asset.kernel.model.DDMStructureClassType;
+import com.liferay.asset.model.DDMStructureClassType;
 import com.liferay.dynamic.data.lists.model.DDLRecordSet;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetServiceUtil;
 import com.liferay.portal.kernel.exception.PortalException;
